+++ src/execution/plugins/script-execution-tools/custom-function-async.test.ts.test.ts
@@ -36,7 +36,7 @@ describe('createPromptbookExecutor + custom async function ', () => {
         tools: {
             natural: new MockedEchoNaturalExecutionTools({ isVerbose: true }),
             script: [
-                new JavascriptEvalExecutionTools({
+                new JavascriptExecutionTools({
                     isVerbose: true,
 
                     // Note: [🕎]